+++ next.config.js
@@ -9,6 +9,9 @@ const nextConfig = {
     ignoreBuildErrors: false,
   },
   images: { unoptimized: true },
+  experimental: {
+    instrumentationHook: true,
+  },
 };
 
 module.exports = nextConfig;